+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Unreleased]
 ### Fixed
 - `no_run` attribute is added to the code example in `README.md` to avoid a compile error ([#38]).
+- `Copy` trait bound is removed from `array::Generic` and `single::Generic ([#37]).
 
 ## [0.3.2] - 2021-08-04
 ### Added
@@ -63,6 +64,7 @@
 - Initial version.
 
 [#38]: https://github.com/toku-sa-n/accessor/pull/38
+[#37]: https://github.com/toku-sa-n/accessor/pull/37
 [#34]: https://github.com/toku-sa-n/accessor/pull/34
 [#33]: https://github.com/toku-sa-n/accessor/pull/33
 [#32]: https://github.com/toku-sa-n/accessor/pull/32